✅ fix: #13 初回動作をいったんPuzzleStudio側に集約
